refactor(error-handling): relocate error utility functions
