+++ src/modules/diet/food/infrastructure/api/application/apiFood.ts
@@ -5,8 +5,8 @@ import { type ApiFood } from '~/modules/diet/food/infrastructure/api/domain/apiF
 import { createSupabaseFoodRepository } from '~/modules/diet/food/infrastructure/api/infrastructure/supabase/supabaseFoodRepository'
 import { markSearchAsCached } from '~/modules/search/application/usecases/cachedSearchCrud'
 import { showError } from '~/modules/toast/application/toastManager'
-import { ORIGINAL_ERROR_SYMBOL } from '~/shared/error/errorHandler'
 import { convertApi2Food } from '~/shared/utils/convertApi2Food'
+import { ORIGINAL_ERROR_SYMBOL } from '~/shared/utils/errorUtils'
 import { logging } from '~/shared/utils/logging'
 
 const foodRepository = createSupabaseFoodRepository()
