+++ src/modules/measure/infrastructure/supabase/supabaseBodyMeasureGateway.ts
@@ -6,8 +6,8 @@ import { type BodyMeasureGateway } from '~/modules/measure/domain/measureGateway
 import { SUPABASE_TABLE_BODY_MEASURES } from '~/modules/measure/infrastructure/supabase/constants'
 import { supabaseBodyMeasureMapper } from '~/modules/measure/infrastructure/supabase/supabaseMeasureMapper'
 import { type User } from '~/modules/user/domain/user'
-import { wrapErrorWithStack } from '~/shared/error/errorHandler'
 import { supabase } from '~/shared/supabase/supabase'
+import { wrapErrorWithStack } from '~/shared/utils/errorUtils'
 import { logging } from '~/shared/utils/logging'
 
 export function createSupabaseBodyMeasureGateway(): BodyMeasureGateway {
