+++ src/modules/user/infrastructure/supabase/supabaseUserRepository.ts
@@ -2,11 +2,11 @@ import { type NewUser, type User, userSchema } from '~/modules/user/domain/user'
 import { type UserRepository } from '~/modules/user/domain/userRepository'
 import { SUPABASE_TABLE_USERS } from '~/modules/user/infrastructure/supabase/constants'
 import { subapaseUserMapper } from '~/modules/user/infrastructure/supabase/supabaseUserMapper'
-import { wrapErrorWithStack } from '~/shared/error/errorHandler'
 import {
   registerSubapabaseRealtimeCallback,
   supabase,
 } from '~/shared/supabase/supabase'
+import { wrapErrorWithStack } from '~/shared/utils/errorUtils'
 
 export function createSupabaseUserRepository(): UserRepository {
   return {
